Move getGOPATH method to paths.go
