+++ cmd/ponzu/add.go
@@ -6,9 +6,7 @@ import (
 	"io"
 	"os"
 	"os/exec"
-	"os/user"
 	"path/filepath"
-	"runtime"
 	"strings"
 )
 
@@ -58,31 +56,6 @@ func getAddon(args []string) error {
 		return addError(err)
 	}
 	return nil
-}
-
-// resolve GOPATH. In 1.8 can be default, or custom. A custom GOPATH can
-// also contain multiple paths, in which case 'go get' uses the first
-func getGOPATH() (string, error) {
-	var gopath string
-	gopath = os.Getenv("GOPATH")
-	if gopath == "" {
-		// not set, find the default
-		usr, err := user.Current()
-		if err != nil {
-			return gopath, err
-		}
-		gopath = filepath.Join(usr.HomeDir, "go")
-	} else {
-		// parse out in case of multiple, retain first
-		if runtime.GOOS == "windows" {
-			gopaths := strings.Split(gopath, ";")
-			gopath = gopaths[0]
-		} else {
-			gopaths := strings.Split(gopath, ":")
-			gopath = gopaths[0]
-		}
-	}
-	return gopath, nil
 }
 
 // this is distinct from copyAll() in that files are copied, not moved,
